+++ client/CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Changed Type labels to Categorized As ([#220](https://github.com/project-lux/lux-frontend/issues/220)).
+- Changed the Person and Group selector in advanced search ([#190](https://github.com/project-lux/lux-frontend/issues/190)).
 
 ### Removed